PropertyItem and ProviderCapabilities
