+++ linx/utils/types/shared-types/propertyItem.ts
@@ -1,0 +1,13 @@
+export interface PropertyItem {
+    Color: null | string;
+    GroupName: null;
+    HasItems: boolean;
+    ImagePath: null | string;
+    IsSelected: boolean;
+    OptionID: number;
+    Order: number;
+    PropertyPath: string;
+    Reference: null;
+    Text: string;
+    Value: string;
+  }
