+++ linx/utils/types/shared-types/providerCapabilities.ts
@@ -1,0 +1,8 @@
+export interface ProviderCapabilities {
+    CanFacet: boolean;
+    CanPage: boolean;
+    CanSort: boolean;
+    CanSpell: boolean;
+    CanTerm: boolean;
+    Provider: string;
+  }
