+++ linx/utils/types/suggestionsJSON.ts
@@ -2,9 +2,11 @@ import { LinxError } from "./common.ts";
 import { CategoryItem } from "./shared-types/categoryItem.ts";
 import { LinxMetadata } from "./shared-types/linxMetada.ts";
 import { NavigationInfo } from "./shared-types/navigationInfo.ts";
+import { PropertyItem } from "./shared-types/propertyItem.ts";
 import { ProductSelectionItem } from "./shared-types/productSelectionItem.ts";
 import { ResponseCallBack } from "./shared-types/responseCallBack.ts";
 import { SellerInfo } from "./shared-types/sellerInfo.ts";
+import { ProviderCapabilities } from "./shared-types/providerCapabilities.ts";
 
 export interface WebPage {
   Response: Response;
@@ -231,23 +233,10 @@ export interface ItemOption {
   Label: string;
   PropertyMetadataID: number;
   PropertyName: string;
-  Values: Value[];
+  Values: PropertyItem[];
   VariationLayoutTemplate: string;
 }
 
-export interface Value {
-  Color: null | string;
-  GroupName: null;
-  HasItems: boolean;
-  ImagePath: null | string;
-  IsSelected: boolean;
-  OptionID: number;
-  Order: number;
-  PropertyPath: string;
-  Reference: null;
-  Text: string;
-  Value: string;
-}
 
 export interface MediaGroup {
   Custom: null;
@@ -304,14 +293,6 @@ export interface Tag {
   TagID: number;
 }
 
-export interface ProviderCapabilities {
-  CanFacet: boolean;
-  CanPage: boolean;
-  CanSort: boolean;
-  CanSpell: boolean;
-  CanTerm: boolean;
-  Provider: string;
-}
 
 export interface SortOption {
   Alias: string;
